func-tests: new testenv module
