+++ functional-tests/fn_bridge_deposit_reclaim.py
@@ -7,8 +7,8 @@ from strata_utils import (
     take_back_transaction,
 )
 
+import testenv
 from constants import DEFAULT_ROLLUP_PARAMS, DEFAULT_TAKEBACK_TIMEOUT, UNSPENDABLE_ADDRESS
-from entry import TestStrata
 from utils import get_bridge_pubkey, wait_until
 
 # Local constants
@@ -19,7 +19,7 @@ TAKE_BACK_FEE = 17_243
 
 
 @flexitest.register
-class BridgeDepositReclaimTest(TestStrata):
+class BridgeDepositReclaimTest(testenv.StrataTester):
     """
     A test class for reclaim path scenarios of bridge deposits.
