+++ functional-tests/fn_client_restart.py
@@ -1,10 +1,10 @@
 import flexitest
 
 import net_settings
+import testenv
 from constants import (
     ERROR_PROOF_ALREADY_CREATED,
 )
-from entry import BasicEnvConfig, TestStrata
 from utils import (
     check_nth_checkpoint_finalized,
     check_submit_proof_fails_for_nonexistent_batch,
@@ -13,11 +13,13 @@ from utils import (
 
 
 @flexitest.register
-class BlockFinalizationSeqRestartTest(TestStrata):
+class BlockFinalizationSeqRestartTest(testenv.StrataTester):
     """This tests finalization when sequencer client restarts"""
 
     def __init__(self, ctx: flexitest.InitContext):
-        ctx.set_env(BasicEnvConfig(101, rollup_settings=net_settings.get_fast_batch_settings()))
+        ctx.set_env(
+            testenv.BasicEnvConfig(101, rollup_settings=net_settings.get_fast_batch_settings())
+        )
 
     def main(self, ctx: flexitest.RunContext):
         seq = ctx.get_service("sequencer")
