+++ functional-tests/fn_sync_l1_finalization_timeout.py
@@ -1,10 +1,10 @@
 import flexitest
 
 import net_settings
+import testenv
 from constants import (
     ERROR_PROOF_ALREADY_CREATED,
 )
-from entry import BasicEnvConfig, TestStrata
 from utils import (
     check_nth_checkpoint_finalized,
     check_submit_proof_fails_for_nonexistent_batch,
@@ -12,7 +12,7 @@ from utils import (
 
 
 @flexitest.register
-class BlockFinalizationTimeoutTest(TestStrata):
+class BlockFinalizationTimeoutTest(testenv.StrataTester):
     """
     This checks for finalization if proof is not submitted within a timeout period
     """
@@ -24,7 +24,7 @@ class BlockFinalizationTimeoutTest(TestStrata):
         settings.genesis_trigger = premine_blocks + 5
         settings.proof_timeout = self.proof_timeout
 
-        ctx.set_env(BasicEnvConfig(premine_blocks, rollup_settings=settings))
+        ctx.set_env(testenv.BasicEnvConfig(premine_blocks, rollup_settings=settings))
 
     def main(self, ctx: flexitest.RunContext):
         seq = ctx.get_service("sequencer")
